Unify permission/guard error handling into withGuardedMutation
Fold withPermissionCheck, GuardError/isWriteConflict, and
Programs.newProgram's duck-typed try/catch into one helper in
guardedMutation.ts, replacing permissionError.ts and transaction.ts.

withGuardedMutation catches GuardError, a write conflict, then the
not-found-under-permission-where case (only when opts.entity/message
are given), falling back to the raw error message. Programs.editSuperUser
and Users.changeRole now call it around their $transaction blocks instead
of hand-rolling the catch, and Programs.newProgram goes through it too,
fixing its duck-typed catch. All other withPermissionCheck call sites
renamed mechanically.

Closes #165
