+++ src/lib/server/actions/Graphs.ts
@@ -8,7 +8,7 @@ import {
 	whereHasGraphCoursePermission,
 	whereHasGraphSandboxPermission
 } from '../permissions';
-import { withPermissionCheck } from './permissionError';
+import { withGuardedMutation } from './guardedMutation';
 import { GraphValidator } from '$lib/validators/graphValidator';
 
 import type {
@@ -118,7 +118,7 @@ export class GraphActions {
 	) {
 		if (!form.valid) return setError(form, '', 'Invalid node positions');
 
-		return await withPermissionCheck(
+		return await withGuardedMutation(
 			() =>
 				prisma.graph.update({
 					where: {
@@ -184,7 +184,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Course',
 				message:
 					'You are not allowed to edit this course. You are not an program admin/editor or course admin/editor'
@@ -205,7 +205,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Sandbox',
 				message: 'You are not allowed to edit this sandbox. You are not an owner or editor'
 			});
@@ -240,7 +240,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Course',
 				message:
 					'You are not allowed to edit this course. You are not an program admin/editor or course admin/editor'
@@ -261,7 +261,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Sandbox',
 				message: 'You are not allowed to edit this sandbox. You are not an owner or editor'
 			});
@@ -293,7 +293,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Course',
 				message:
 					'You are not allowed to edit this course. You are not an program admin/editor or course admin/editor'
@@ -311,7 +311,7 @@ export class GraphActions {
 				}
 			});
 
-			return await withPermissionCheck(() => query, form, 'name', {
+			return await withGuardedMutation(() => query, form, 'name', {
 				entity: 'Sandbox',
 				message: 'You are not allowed to edit this sandbox. You are not an owner or editor'
 			});
